--- a/crates/zk-core/src/check.rs
+++ b/crates/zk-core/src/check.rs
@@ -1,15 +1,8 @@
 //! QuickSilver consistency check.
 
-use std::mem;
-
 use blake3::Hasher;
-use cfg_if::cfg_if;
-use mpz_core::{
-    Block,
-    bitvec::{BitSlice, BitVec},
-};
+use mpz_core::Block;
 use serde::{Deserialize, Serialize};
-use zerocopy::IntoBytes;
 
 use crate::vole::{vole_receiver, vole_sender};
 
@@ -29,185 +22,180 @@ pub(crate) struct Triple {
     pub(crate) z: Block,
 }
 
+/// Prover’s unblinded preprocessed values for the consistency check.
+///
+/// These are random linear combinations aggregated over one or more
+/// circuit executions. The values `u` and `v` are kept private and will later
+/// be masked (blinded) with random elements to derive the public values
+/// `U = u + u*` and `V = v + v*` sent to the verifier.
 #[derive(Debug, Default)]
-pub(crate) struct Check {
-    triples: Vec<Triple>,
-    adjust: BitVec,
+pub(crate) struct ProverCheck {
+    u: Option<Block>,
+    v: Option<Block>,
 }
 
-impl Check {
-    /// Reserves capacity for at least `n` AND gates, returns the starting
-    /// index.
-    pub(crate) fn reserve(&mut self, n: usize) -> usize {
-        let idx = self.triples.len();
-        self.triples.resize_with(idx + n, Default::default);
-        self.adjust.resize_with(idx + n, |_| Default::default());
-        idx
-    }
-
-    pub(crate) fn write(&mut self, idx: usize, triples: &[Triple], adjust: &BitSlice) {
-        self.triples[idx..idx + triples.len()].copy_from_slice(triples);
-        self.adjust[idx..idx + triples.len()].copy_from_bitslice(adjust);
-    }
-
-    /// Returns `true` if there are gates to check.
-    #[inline]
-    pub(crate) fn wants_check(&self) -> bool {
-        !self.triples.is_empty()
-    }
-
-    fn compute_chis(&self, mut chi: Block) -> Vec<Block> {
-        // TODO: Consider using a PRG instead so computing the coefficients
-        // can be done in parallel.
-        let mut chis = Vec::with_capacity(self.triples.len());
-        chis.push(chi);
-        for _ in 1..self.triples.len() {
-            chi = chi.gfmul(chi);
-            chis.push(chi);
+impl ProverCheck {
+    /// Aggregates the values for the consistency check.
+    pub fn update(&mut self, u: Block, v: Block) {
+        match &mut self.u {
+            Some(prev_u) => *prev_u ^= u,
+            None => self.u = Some(u),
         }
 
-        chis
+        match &mut self.v {
+            Some(prev_v) => *prev_v ^= v,
+            None => self.v = Some(v),
+        }
     }
 
     /// Executes the prover check, returning `U` and `V` defined in Figure 5,
     /// Step 7.b.
-    pub(crate) fn check_prover(
-        &mut self,
-        transcript: &mut Hasher,
-        svole_choices: &[bool],
-        svole_ev: &[Block],
-    ) -> Result<UV> {
-        #[inline]
-        fn compute_terms(triple: Triple, chi: Block) -> (Block, Block) {
-            let Triple { x, y, z } = triple;
-
-            let u = x.gfmul(y).gfmul(chi);
-
-            // (Note that the LSB of a MAC contains the authenticated bit).
-            let a_10 = if x.lsb() { y } else { Block::ZERO };
-            let a_11 = if y.lsb() { x } else { Block::ZERO };
-            let v = (a_10 ^ a_11 ^ z).gfmul(chi);
-
-            (u, v)
-        }
-
-        let adjust_len = self.adjust.len();
-        transcript.update(&self.adjust.as_raw_slice().as_bytes()[..adjust_len.div_ceil(8)]);
-
-        let chi = Block::try_from(&transcript.finalize().as_bytes()[..16])
-            .expect("block should be 16 bytes");
-        let chis = self.compute_chis(chi);
-        let macs = mem::take(&mut self.triples);
-        cfg_if! {
-            if #[cfg(all(feature = "rayon", not(feature = "force-st")))] {
-                use rayon::prelude::*;
-
-                let (mut u, mut v) = macs
-                    .into_par_iter()
-                    .zip(chis)
-                    .map(|(macs, chi)| compute_terms(macs, chi))
-                    .reduce(
-                        || (Block::ZERO, Block::ZERO),
-                        |(u_acc, v_acc), (u, v)| (u_acc ^ u, v_acc ^ v),
-                    );
-            } else {
-                let (mut u, mut v) = macs
-                    .into_iter()
-                    .zip(chis)
-                    .map(|(macs, chi)| compute_terms(macs, chi))
-                    .fold(
-                        (Block::ZERO, Block::ZERO),
-                        |(u_acc, v_acc), (u, v)| (u_acc ^ u, v_acc ^ v),
-                    );
-            }
-        }
-
+    pub fn check(&mut self, svole_choices: &[bool], svole_ev: &[Block]) -> Result<UV> {
         let (a_0, a_1) = vole_receiver(
             svole_choices.try_into().map_err(|_| CheckError::SVole)?,
             svole_ev.try_into().map_err(|_| CheckError::SVole)?,
         );
 
-        u ^= a_0;
-        v ^= a_1;
+        let u = self.u.ok_or(CheckError::NoExecutions)?;
+        let v = self.v.ok_or(CheckError::NoExecutions)?;
 
-        transcript.update(&u.to_bytes());
-        transcript.update(&v.to_bytes());
+        // Reset the values.
+        self.u = None;
+        self.v = None;
 
-        self.adjust.clear();
+        Ok(UV {
+            u: u ^ a_0,
+            v: v ^ a_1,
+        })
+    }
 
-        Ok(UV { u, v })
+    /// Returns `true` if there are gates to check.
+    #[inline]
+    pub(crate) fn wants_check(&self) -> bool {
+        // `true` if at least one AND gate was executed.
+        self.u.is_some()
+    }
+}
+
+/// Verifier’s unmasked preprocessed values for the consistency check.
+///
+/// This is a random linear combination aggregated over one or more
+/// circuit executions. The value will later be masked with a random
+/// element to derive `W`.
+#[derive(Debug, Default)]
+pub(crate) struct VerifierCheck {
+    w: Option<Block>,
+}
+
+impl VerifierCheck {
+    /// Aggregates the values for the consistency check.
+    pub fn update(&mut self, w: Block) {
+        match &mut self.w {
+            Some(prev_w) => *prev_w ^= w,
+            None => self.w = Some(w),
+        }
     }
 
     /// Executes the verifier check, returning `W` defined in Figure 5, Step
     /// 7.c.
-    pub(crate) fn check_verifier(
-        &mut self,
-        transcript: &mut Hasher,
-        delta: &Block,
-        svole_keys: &[Block],
-        uv: UV,
-    ) -> Result<()> {
-        #[inline]
-        fn compute_term(triple: Triple, chi: Block, delta: &Block) -> Block {
-            let Triple { x, y, z } = triple;
-            let b = x.gfmul(y) ^ delta.gfmul(z);
-            b.gfmul(chi)
-        }
+    pub fn check(&mut self, delta: &Block, svole_keys: &[Block], uv: UV) -> Result<()> {
+        let w = self.w.ok_or(CheckError::NoExecutions)?;
 
-        let adjust_len = self.adjust.len();
-        transcript.update(&self.adjust.as_raw_slice().as_bytes()[..adjust_len.div_ceil(8)]);
-
-        let chi = Block::try_from(&transcript.finalize().as_bytes()[..16])
-            .expect("block should be 16 bytes");
-        let chis = self.compute_chis(chi);
-        let keys = mem::take(&mut self.triples);
-        cfg_if! {
-            if #[cfg(all(feature = "rayon", not(feature = "force-st")))] {
-                use rayon::prelude::*;
-
-                let mut w = keys
-                    .into_par_iter()
-                    .zip(chis)
-                    .map(|(keys, chi)| compute_term(keys, chi, delta))
-                    .reduce(
-                        || Block::ZERO,
-                        |w_acc, w| w_acc ^ w,
-                    );
-            } else {
-                let mut w = keys
-                    .into_iter()
-                    .zip(chis)
-                    .map(|(keys, chi)| compute_term(keys, chi, delta))
-                    .fold(
-                        Block::ZERO,
-                        |w_acc, w| w_acc ^ w,
-                    );
-            }
-        }
-
-        let b = vole_sender(svole_keys.try_into().map_err(|_| CheckError::SVole)?);
-
-        w ^= b;
+        let b = vole_sender(
+            &svole_keys
+                .try_into()
+                .map_err(|_| CheckError::SVole)
+                .unwrap(),
+        );
 
         let UV { u, v } = uv;
-        transcript.update(&u.to_bytes());
-        transcript.update(&v.to_bytes());
 
-        self.adjust.clear();
-
-        if w != u ^ delta.gfmul(v) {
-            // Invalid! Call the police.
+        if w ^ b != u ^ delta.gfmul(v) {
             return Err(CheckError::Invalid);
         }
 
+        // Reset the value.
+        self.w = None;
+
         Ok(())
+    }
+
+    /// Returns `true` if there are gates to check.
+    #[inline]
+    pub(crate) fn wants_check(&self) -> bool {
+        // `true` if at least one AND gate was executed.
+        self.w.is_some()
     }
 }
 
+/// Computes the values `u` and `v` as defined in [ProverCheck].
+pub(crate) fn compute_uv(transcript: &mut Hasher, triples: &[Triple]) -> (Block, Block) {
+    #[inline]
+    fn compute_terms(triple: Triple, chi: Block) -> (Block, Block) {
+        let Triple { x, y, z } = triple;
+
+        let u = x.gfmul(y).gfmul(chi);
+
+        // (Note that the LSB of a MAC contains the authenticated bit).
+        let a_10 = if x.lsb() { y } else { Block::ZERO };
+        let a_11 = if y.lsb() { x } else { Block::ZERO };
+        let v = (a_10 ^ a_11 ^ z).gfmul(chi);
+
+        (u, v)
+    }
+
+    let chi =
+        Block::try_from(&transcript.finalize().as_bytes()[..16]).expect("block should be 16 bytes");
+    let chis = compute_chis(chi, triples.len());
+
+    triples
+        .iter()
+        .zip(chis)
+        .map(|(mac, chi)| compute_terms(*mac, chi))
+        .fold((Block::ZERO, Block::ZERO), |(u_acc, v_acc), (u, v)| {
+            (u_acc ^ u, v_acc ^ v)
+        })
+}
+
+/// Computes the value `w` as defined in [VerifierCheck].
+pub(crate) fn compute_w(transcript: &mut Hasher, triples: &[Triple], delta: &Block) -> Block {
+    #[inline]
+    fn compute_term(triple: Triple, chi: Block, delta: &Block) -> Block {
+        let Triple { x, y, z } = triple;
+        let b = x.gfmul(y) ^ delta.gfmul(z);
+        b.gfmul(chi)
+    }
+
+    let chi =
+        Block::try_from(&transcript.finalize().as_bytes()[..16]).expect("block should be 16 bytes");
+    let chis = compute_chis(chi, triples.len());
+
+    triples
+        .iter()
+        .zip(chis)
+        .map(|(key, chi)| compute_term(*key, chi, delta))
+        .fold(Block::ZERO, |w_acc, w| w_acc ^ w)
+}
+
+fn compute_chis(mut chi: Block, count: usize) -> Vec<Block> {
+    debug_assert!(count > 0);
+
+    let mut chis = Vec::with_capacity(count);
+    chis.push(chi);
+    for _ in 1..count {
+        chi = chi.gfmul(chi);
+        chis.push(chi);
+    }
+
+    chis
+}
+
 #[derive(Debug, thiserror::Error)]
-pub(crate) enum CheckError {
+pub enum CheckError {
     #[error("incorrect number of sVOLE instances provided")]
     SVole,
     #[error("invalid consistency check")]
     Invalid,
+    #[error("check was called when no AND gates were executed")]
+    NoExecutions,
 }