… crates/zk/Cargo.toml
… crates/zk/src/prover.rs
… crates/zk/src/verifier.rs