This PR improves QuickSilver performance by ~ 50%. Notable changes:
- preprocesses the consistency check values immediately after a circuit execution, eliminating the need to store triples in memory (and later copying them)

- stops using rayon for preprocessing the consistency check values, to avoid starving the threads executing the circuits

- performs a separate consistency check for each circuit


### Benchmarks
running `cargo bench` in mpz/crates/zk/benches 

before this PR
`zk/aes128               time:   [178.00 ms 187.11 ms 198.20 ms]`

after this PR
`zk/aes128               time:   [100.36 ms 103.20 ms 106.35 ms]`